+++ database/migrations/14-v_deviation_trees_gavle.sql
@@ -39,8 +39,8 @@ ALTER TABLE upstream.mv_match_trees_gavle OWNER TO app;
 
 CREATE OR REPLACE VIEW upstream.v_deviation_trees_gavle AS
 	SELECT
-		5 AS dataset_id,
-		16 AS layer_id,
+		5::bigint AS dataset_id,
+		16::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
